% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get_areas.R
\name{area_phases_as_df}
\alias{area_phases_as_df}
\title{Convert areas phase list to data frame.}
\usage{
area_phases_as_df(phases_list)
}
\arguments{
\item{phases_list}{Phases list}
}
\description{
Cleans up the phases list to allow unnesting. Drops the \code{color}
column, then converts to data frame.
}
