% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_utils.R
\name{ensure_list}
\alias{ensure_list}
\title{Ensure returned value is a list}
\usage{
ensure_list(ret)
}
\description{
Sometimes if parameters are passed, the returned value is not a list of lists,
but is instead a data frame. If that is the case, we need to return it nested
within a list for parsing later.
}
