% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_download.R
\name{ipc_download}
\alias{ipc_download}
\title{Download point IPC data from the API}
\usage{
ipc_download(wrangle = TRUE, configuration = NULL)
}
\arguments{
\item{wrangle}{If \code{True} (the default), wrangles the data into a coherent
data structure. Details on the wrangling in the documentation of
\code{\link[=ipc_wrangle]{ipc_wrangle()}}.}

\item{configuration}{Configuration to pass on to \code{\link[rhdx:pull_dataset]{rhdx::pull_dataset()}}.}
}
\value{
IPC data frame.
}
\description{
\code{ipc_get()} downloads IPC country data from the \href{https://data.humdata.org/dataset/ipc-country-data}{Humanitarian Data Exchange}.
}
