% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get_country.R
\name{ipc_get_country}
\alias{ipc_get_country}
\title{Get country resources from the IPC API}
\usage{
ipc_get_country(country = NULL, year = NULL, type = NULL, api_key = NULL)
}
\arguments{
\item{country}{ISO2 country code.}

\item{year}{Single numeric year to filter analysis, calculated from the
maximum year of current period dates. If \code{NULL}, the default, returns
data for all years.}

\item{type}{Single string value of 'A' or 'C', corresponding to food security
conditions, either acute or chronic. If \code{NULL}, the default, returns
data for all types.}

\item{api_key}{IPC API key. If \code{NULL} (the default), looks for \code{IPC_API_KEY}
in the environment.}
}
\value{
A data frame.
}
\description{
Accesses the country resources on the IPC API. Contains detailed
country-level data. If \code{country}, \code{year} and/or \code{type} parameters are passed,
accesses the \strong{country} public API endpoint and pulls in all country data
filtered to those parameters.
}
\details{
Country data is the highest level of aggregation for IPC analysis. Data is
the population estimates and other details aggregated from area and point
analyses within the country for that specific analysis. No phase
classifications are made at the country level. Data is currently output only
for current periods through the \strong{country} endpoint. Use
\code{\link[=ipc_get_population]{ipc_get_population()}} to get detailed population data for all analysis
periods.

See the \href{https://www.ipcinfo.org}{IPC website} and
\href{https://docs.api.ipcinfo.org}{API documentation} for more information.
}
\examples{
\dontrun{
# get all areas from the public API
ipc_get_country()

# get country data just for Somalia
ipc_get_country(country = "SO")
}

}
