% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get_population.R
\name{ipc_get_population}
\alias{ipc_get_population}
\title{Get population resources from the IPC API}
\usage{
ipc_get_population(
  country = NULL,
  start = NULL,
  end = NULL,
  id = NULL,
  api_key = NULL
)
}
\arguments{
\item{country}{ISO2 country code.}

\item{start}{Start year.}

\item{end}{End year.}

\item{id}{Analysis ID.}

\item{api_key}{IPC API key. If \code{NULL} (the default), looks for \code{IPC_API_KEY}
in the environment.}
}
\value{
A list of 3 data frames:
\itemize{
\item Country data frame
\item Groups data frame
\item Areas data frame
}
}
\description{
Accesses the population resources on the IPC API. Contains detailed
population data. If \code{country} and/or \code{start} and \code{end} parameters are passed,
accesses the \strong{population} developer API endpoint and pulls in all data.
filtered by those parameters. To get the population data for a specific
analysis, available on the \strong{types/{id}} developer API endpoint,
pass in \code{id}. You cannot pass in both sets of parameters.
}
\details{
Unlike the other \code{ipc_get_..()} family of functions, this returns a list of
datasets, corresponding to \code{country} data, \code{groups} data, and
\code{areas}/\code{points}data. The benefit of \code{ipc_get_population()} is that the
returned data for each level of analysis contains all periods of analysis.
Areas and points data is the lowest level of IPC analysis where population
estimates for each phase are provided and a general area-level classification
is made. Groups data, where available, are geographies within a country that
comprise multiple areas and/or points. There is no phase classification at
the group level, but populations in each phase are provided. The same applies
to country-level data.

These datasets are available elsewhere through:
\itemize{
\item Country data: \code{\link[=ipc_get_country]{ipc_get_country()}}
\item Groups data: Not available through other functions
\item Areas and points data: \code{\link[=ipc_get_areas]{ipc_get_areas()}}
}

See the respective function documentation for more details on what each
dataset comprises or the \href{https://www.ipcinfo.org}{IPC website} and
\href{https://docs.api.ipcinfo.org}{API documentation} for more detailed and
comprehensive information on the data and analysis.
}
\examples{
\dontrun{
# get all areas from the public API
ipc_get_population()

# get areas for specific analysis ID and period from developer API
ipc_get_population(id = 12856213)
}

}
