% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_wrangle.R
\name{ipc_wrangle}
\alias{ipc_wrangle}
\title{Wrangle IPC data}
\usage{
ipc_wrangle(df)
}
\arguments{
\item{df}{Raw IPC data returned from \code{ipc_download(wrangle = FALSE)}.}
}
\value{
Wrangled IPC data.
}
\description{
\code{ipc_wrangle()} transforms the IPC data \href{https://data.humdata.org/dataset/ipc-country-data}{stored on HDX}. The outcome is a
tidy dataset where each row corresponds to an analytical output (e.g. first
projection for a specific date) and geographical area. Performed by default
in \code{\link[=ipc_download]{ipc_download()}}.
}
\details{
There are a variety of transformations done in \code{ipc_wrangle()}.
\subsection{Renaming}{

Column names are extracted from the dataset and then cleaned up before
applying them to the dataframe. Special characters are converted to
alphanumeric and strings converted to lower case.
}

\subsection{Area classifications}{

Area and country classifications are mixed in the raw output dataset. In some
cases, the country column contains the specific area being classified. This
is usually when an analysis has been done without area level phase
classification, and is seen in Yemen, South Sudan and particularly Somalia
analyses. This code detects these cases and extracts the correct country and
area names.
}

\subsection{Pivot}{

Data is pivoted so that instead of specific columns based on the analytical
type (i.e. current, first projection, or second projection), they are stored
in individual rows.
}

\subsection{Final cleaning}{

Extraneous columns are dropped, as are rows without data or those that are
aggregated from component rows (to fit the tidy principle of one level of
analysis being in the rows). Additional columns are produced from the
analysis period to specify in date format the start and end dates of the
analysis period, useful for future analysis or plotting.
}
}
