% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get_points.R
\name{point_phases_as_df}
\alias{point_phases_as_df}
\title{Convert phases list to data frame.}
\usage{
point_phases_as_df(phases_list)
}
\arguments{
\item{phases_list}{Phases list}
}
\description{
Cleans up the phases list to allow unnesting. Drops the \code{overall_phase}
column, then converts to data frame.
}
