% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ripc-package.R
\docType{package}
\name{Ripc-package}
\alias{Ripc}
\alias{Ripc-package}
\title{Ripc: Download and Process IPC Data}
\description{
Utilities to access Integrated Food Security Phase Classification (IPC) and Cadre Harmonisé (CH) food security data. Wrapper functions are available for all of the IPC-CH Public API simplified and advanced endpoints to easily download the data in a clean and tidy format.
}
\author{
\strong{Maintainer}: Seth Caldwell \email{caldwellst@gmail.com}

}
\keyword{internal}
