% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get.R
\name{ipc_get}
\alias{ipc_get}
\title{Get resource from IPC API}
\usage{
ipc_get(resource, api_key = NULL, ..., drop_metadata = FALSE)
}
\arguments{
\item{resource}{One of the resources exposed in the IPC API, such
\code{"areas"} , \code{"analyses"}, \code{"points"}, or \code{"country"}.}

\item{api_key}{IPC API key. If \code{NULL} (the default), looks for \code{IPC_API_KEY}
in the environment.}

\item{...}{Named parameters passed to the API call URL in the form of
\code{argument=value}.}

\item{drop_metadata}{Logical, whether or not to drop metadata from the
original list return.}
}
\value{
Data frame from the API
}
\description{
Back-end function used to drive the API calls of the other \code{ipc_get_...()}
family of functions.
}
