% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get.R
\name{ipc_get}
\alias{ipc_get}
\title{Get resource from IPC API}
\usage{
ipc_get(resource, return_format, pass_format, api_key = NULL, ...)
}
\arguments{
\item{resource}{One of the resources exposed in the IPC API, such
\code{"areas"} , \code{"analyses"}, \code{"points"}, or \code{"country"}.}

\item{return_format}{The format that should be returned by the API. Allows
\code{"csv"} (the default) and \code{"geojson"}.}

\item{pass_format}{Pass format explicitly in the API call.}

\item{api_key}{IPC API key. If \code{NULL} (the default), looks for \code{IPC_API_KEY}
in the environment.}

\item{...}{Named parameters passed to the API call URL in the form of
\code{argument=value}.}
}
\value{
Data frame from the API
}
\description{
Back-end function used to drive the API calls of the other \code{ipc_get_...()}
family of functions.
}
