% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_get_analyses.R
\name{ipc_get_analyses}
\alias{ipc_get_analyses}
\title{Get analyses resource from IPC API}
\usage{
ipc_get_analyses(country = NULL, year = NULL, type = NULL, api_key = NULL)
}
\arguments{
\item{country}{ISO2 country code.}

\item{year}{Single numeric year to filter analysis, calculated from the
maximum year of current period dates. If \code{NULL}, the default, returns
data for all years.}

\item{type}{Single string value of 'A' or 'C', corresponding to food security
conditions, either acute or chronic. If \code{NULL}, the default, returns
data for all types.}

\item{api_key}{IPC API key. If \code{NULL} (the default), looks for \code{IPC_API_KEY}
in the environment.}
}
\description{
Accesses the analyses resource on the IPC API. More details to come.
}
