% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ripc-package.R
\docType{package}
\name{Ripc-package}
\alias{Ripc}
\alias{Ripc-package}
\title{Ripc: Download and Process IPC Data}
\description{
Download and wrangling functionality for IPC data stored on the Humanitarian Data Exchange.
}
\author{
\strong{Maintainer}: Seth Caldwell \email{caldwellst@gmail.com}

}
\keyword{internal}
