% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ripc-package.R
\docType{package}
\name{ripc-package}
\alias{ripc}
\alias{ripc-package}
\title{ripc: Download and Tidy IPC and CH Data}
\description{
Utilities to access Integrated Food Security Phase Classification (IPC) and Cadre Harmonisé (CH) food security data. Wrapper functions are available for all of the 'IPC-CH' Public API (\url{https://docs.api.ipcinfo.org}) simplified and advanced endpoints to easily download the data in a clean and tidy format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocha-dap/ripc}
  \item Report bugs at \url{https://github.com/ocha-dap/ripc/issues}
}

}
\author{
\strong{Maintainer}: Giulia Martini \email{giulia.martini@un.org} [copyright holder]

}
\keyword{internal}
